+++ packages/patternfly-3/react-console/src/DesktopViewer/ManualConnection.js
@@ -2,7 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import consoleDetailPropType from './consoleDetailPropType';
-import { Form, FormGroup, Col } from 'patternfly-react';
+import { Form, Grid } from 'patternfly-react';
+
+const { Col } = Grid;
+const { FormGroup } = Form;
 
 const Detail = ({ title, value }) => (
   <FormGroup>
@@ -22,6 +25,7 @@ Detail.propTypes = {
 const ManualConnection = ({
   spice,
   vnc,
+  rdp,
   textSpice,
   textVNC,
   textNoProtocol,
@@ -33,10 +37,13 @@ const ManualConnection = ({
   textSpicePort,
   textVNCPort,
   textSpiceTlsPort,
-  textVNCTlsPort
+  textVNCTlsPort,
+  textRDPPort,
+  textRdpAddress
 }) => {
   const msg = spice || vnc ? textConnectWith : textNoProtocol;
   const address = spice && vnc && spice.address === vnc.address && spice.address; // merge value if equal
+  const rdpAddress = rdp && rdp.address !== address ? rdp.address : null;
 
   return (
     <div className="manual-connection-pf">
@@ -45,11 +52,13 @@ const ManualConnection = ({
       <Form horizontal>
         {address && <Detail title={textAddress} value={address} />}
         {!address && spice && <Detail title={textSpiceAddress} value={spice.address} />}
+        {rdpAddress && <Detail title={textRdpAddress} value={rdpAddress} />}
         {spice && spice.port && <Detail title={textSpicePort} value={spice.port} />}
         {spice && spice.tlsPort && <Detail title={textSpiceTlsPort} value={spice.tlsPort} />}
         {!address && vnc && <Detail title={textVNCAddress} value={vnc.address} />}
         {spice && vnc.port && <Detail title={textVNCPort} value={vnc.port} />}
         {spice && vnc.tlsPort && <Detail title={textVNCTlsPort} value={vnc.tlsPort} />}
+        {rdp && rdp.port && <Detail title={textRDPPort} value={rdp.port} />}
       </Form>
     </div>
   );
@@ -58,6 +67,7 @@ const ManualConnection = ({
 ManualConnection.propTypes = {
   spice: consoleDetailPropType,
   vnc: consoleDetailPropType,
+  rdp: consoleDetailPropType,
 
   textManualConnection: PropTypes.string.isRequired /** Internationalization */,
   textSpice: PropTypes.string.isRequired,
@@ -71,12 +81,15 @@ ManualConnection.propTypes = {
   textSpicePort: PropTypes.string.isRequired,
   textVNCPort: PropTypes.string.isRequired,
   textSpiceTlsPort: PropTypes.string.isRequired,
-  textVNCTlsPort: PropTypes.string.isRequired
+  textVNCTlsPort: PropTypes.string.isRequired,
+  textRDPPort: PropTypes.string.isRequired,
+  textRdpAddress: PropTypes.string.isRequired
 };
 
 ManualConnection.defaultProps = {
   spice: null,
-  vnc: null
+  vnc: null,
+  rdp: null
 };
 
 export default ManualConnection;