+++ packages/patternfly-3/react-console/src/common/constants.js
@@ -1,4 +1,37 @@
-export const NONE_TYPE = '_none_';
-export const SERIAL_CONSOLE_TYPE = 'SerialConsole';
-export const SPICE_CONSOLE_TYPE = 'SpiceConsole';
-export const VNC_CONSOLE_TYPE = 'VncConsole';
+const NONE_TYPE = '_none_';
+const SERIAL_CONSOLE_TYPE = 'SerialConsole';
+const SPICE_CONSOLE_TYPE = 'SpiceConsole';
+const VNC_CONSOLE_TYPE = 'VncConsole';
+const RDP_CONSOLE_TYPE = 'RdpConsole';
+
+const CONNECTING = 'connecting';
+const CONNECTED = 'connected';
+const DISCONNECTED = 'disconnected';
+const LOADING = 'loading';
+
+const DEFAULT_VV_FILENAME = 'console.vv';
+const DEFAULT_VV_MIMETYPE = 'application/x-virt-viewer';
+const DEFAULT_RDP_FILENAME = 'console.rdp';
+const DEFAULT_RDP_MIMETYPE = 'application/rdp';
+const DEFAULT_RDP_PORT = 3389;
+
+const constants = {
+  NONE_TYPE,
+  SERIAL_CONSOLE_TYPE,
+  SPICE_CONSOLE_TYPE,
+  VNC_CONSOLE_TYPE,
+  RDP_CONSOLE_TYPE,
+
+  CONNECTING,
+  CONNECTED,
+  DISCONNECTED,
+  LOADING,
+
+  DEFAULT_VV_FILENAME,
+  DEFAULT_VV_MIMETYPE,
+  DEFAULT_RDP_FILENAME,
+  DEFAULT_RDP_MIMETYPE,
+  DEFAULT_RDP_PORT
+};
+
+export default constants;