+++ packages/patternfly-4/react-docs/src/components/navigation/navigation.styles.js
@@ -27,4 +27,5 @@ export default StyleSheet.create({
   search: {
     margin: spacerMd.var
   }
+  }
 });